fixes to lspconfig and null-ls.
Attempting to get goimports to run on save. Not sure if null-ls is
loading (I think it is, but I strangely don't have access to commands
like :NullLsInfo). I fixed issues with the nvim-treesitter configuration
(missing commas).

Rewrote on_attach function and placed it in "custom.configs.lsp.utils",
I also passed it to null_ls.setup, which I'm not sure is necessary.
It didn't seem to have an effect

All in all, at a good place. The main persisting issue is that go imports
don't get automatically removed on save when there is no remaining
references. I believe the issue is with my null-ls setup.

I would like to implement the go.nvim package, but removing it seemed to
solve a lot of my problems.