Get rid of auto-session. also removed workspaces
I'm sick of auto-session. Not working the way I need it to, just end up
with a million buffers.

I also removed the workspaces plugin, I can't even remember what it
does. Only removed it because I thought that it was the plugin that was
maintaining sessions. We'll see if it pops up as a problem later on.