modular tmux config for machine-specific settings
Tmux config language does not allow conditionals, so I'm creating a
minimal config that can be shared among all computers, then an
additional architecture specific config (although this is also dependent
on the terminal-tmux interaction e.g., using iTerm with tmux on an x86 Mac would
not necessarily cause the same problems as using Alacritty with tmux on
an x86 Mac). We are then using a script to create the actual `.tmux.conf`
that is sourced by tmux.

This config consists of two statements which source the other two config
files. The ordering is important, as the minimal config, which ends with
a call to initialize the TMUX plugin manager, must come second.

The reason I had to separate these settings is because the way I had
configure Tmux on my M1 Macbook (to get italics and color to work in
nvim) would cause the x86 iMac to break when using Tmux. A single keypress on
the iMac when using Tmux under those conditions would cause the cursor to
jump around and strings of text to be inserted.

Link that provided the fix for the M1 Macbook Air
https://gist.github.com/andersevenrud/015e61af2fd264371032763d4ed965b6