Add indent-blankline plugin, for certain file types
I primarily want this plugin for yaml files where indentation is the
primary structure. It might be useful for python too, but I haven't
started writing much python yet.

The scope feature, which highlights the indentation of the current
lexical scope using the treesitter API, does not integrate well with
alacritty. The underline does not render correctly, tmux may play a role
in this as well. I do not want to open that can of worms again, so I
just turned off the underline attribute in the scope table.

I want to implement a toggle for the indent lines, I have found they are
more distracting than useful when writing Go, so I only have them active
on yaml, html, and json files.

The catppuccin config doesn't seem to be working for integrating the
color scheme, so I set it directly in the plugin, using the hooks
provided.